Move JoinPage to new file.
